refactor(phase-2): remove redundant compat layer files
- Delete 6 compat layer files (engine.py, models.py, ai_services.py,
  config_manager.py, exporters.py, video.py) - all were pure re-exports
- Move SubtitleExporter from exporters.py to its proper home at
  scenefab.services.export.subtitle_exporter
- Move LLMProviderType enum and LLMConfig/CacheConfig dataclasses into
  scenefab.settings (add is_valid() methods to LLMConfig/CacheConfig)
- Update all 7+ import sites (pipeline.py, cli.py, application.py,
  project_manager.py, project_template_manager.py, settings_manager.py,
  direct_video_exporter.py, pipeline_integrator.py) to import from
  specific submodule paths instead of compat layers
- Update test_config_manager.py to use new LLMConfig signature (name=)

Net: -282 lines, 6 files removed
Tests: 46 passed, 1 skipped
Lint: ruff check all green
