+++ src/scenefab/settings_manager.py
@@ -16,7 +16,7 @@ import logging
 
 from scenefab._signals import QObject, Signal
 
-from .config_manager import ConfigManager
+from .settings import ConfigManager
 from .secure_key_manager import get_secure_key_manager
 from .settings_types import SettingType, SettingDefinition, ProjectSettingsProfile
 from .settings_data import get_all_settings_definitions
